Add some missing dependencies to the fabric podspec (#26197)
Summary:
Managed to get fabric building in my app. Here are the changes needed in the podspec.

There were other issues with codegen but it would be too involved to fix for now. This is a good first spec to get it working better in OSS.

## Changelog

[iOS] [Fixed] - Add some missing dependencies to the fabric podspec
Pull Request resolved: https://github.com/facebook/react-native/pull/26197

Test Plan: Managed to get fabric building in a project.

Reviewed By: cpojer

Differential Revision: D17091504

Pulled By: osdnk

fbshipit-source-id: a5d0c46899e342e07ceafc2e4514a03fabf1471d