ReactEventEmitter shouldn't emit events before Catalyst is set up / after it's torn down
Summary:
This should resolve some of the crashes in T9871006. It's unclear if these events are being emitted before Catalyst is set up or after it's torn down, but it's likely a race between teardown and some other event.

Should be safe to just log and continue / fail silently, if this is happening after teardown.

Also upgrading logs to error so that these get logged to logviews.

Changelog: [Internal], will fail silently in a marginal number of existing edge-case crashes for events sent before startup or after shutdown

Reviewed By: mdvacca

Differential Revision: D19269891

fbshipit-source-id: 30fc28693c0be7a15bfc5b5bc7eb88e10c7b4a5a